Add unit tests for Vault Secret Store

+++ pkg/connection/store/vault/store.go
@@ -21,21 +21,21 @@ import (
 	"path/filepath"
 
 	"github.com/hashicorp/vault/api"
-
-	"github.com/crossplane/crossplane-runtime/pkg/resource"
-
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	v1 "github.com/crossplane/crossplane-runtime/apis/common/v1"
 	"github.com/crossplane/crossplane-runtime/pkg/connection/store"
+	kvclient "github.com/crossplane/crossplane-runtime/pkg/connection/store/vault/client"
 	"github.com/crossplane/crossplane-runtime/pkg/errors"
+	"github.com/crossplane/crossplane-runtime/pkg/resource"
 )
 
 // Error strings.
 const (
-	errNoConfig     = "no Vault config provided"
-	errNewClient    = "cannot create new client"
-	errExtractToken = "cannot extract token"
+	errNoConfig        = "no Vault config provided"
+	errNewClient       = "cannot create new client"
+	errExtractToken    = "cannot extract token"
+	errNoTokenProvided = "token auth configured but no token provided"
 
 	errGet    = "cannot get secret"
 	errApply  = "cannot apply secret"
@@ -44,12 +44,12 @@ const (
 
 // KVClient is a Vault AdditiveKVClient Secrets engine client that supports both v1 and v2.
 type KVClient interface {
-	Get(path string, secret *KVSecret) error
-	Apply(path string, secret *KVSecret) error
+	Get(path string, secret *kvclient.KVSecret) error
+	Apply(path string, secret *kvclient.KVSecret) error
 	Delete(path string) error
 }
 
-// SecretStore is a Vault Secret Store.
+// SecretStore is a Vault KVSecret Store.
 type SecretStore struct {
 	client KVClient
 
@@ -72,7 +72,7 @@ func NewSecretStore(ctx context.Context, kube client.Client, cfg v1.SecretStoreC
 	switch cfg.Vault.Auth.Method {
 	case v1.VaultAuthToken:
 		if cfg.Vault.Auth.Token == nil {
-			return nil, errors.New("token auth configured but no token provided")
+			return nil, errors.New(errNoTokenProvided)
 		}
 		t, err := resource.CommonCredentialExtractor(ctx, cfg.Vault.Auth.Token.Source, kube, cfg.Vault.Auth.Token.CommonCredentialSelectors)
 		if err != nil {
@@ -86,15 +86,15 @@ func NewSecretStore(ctx context.Context, kube client.Client, cfg v1.SecretStoreC
 	}
 
 	return &SecretStore{
-		client:            NewAdditiveKVClient(c.Logical(), cfg.Vault.MountPath, WithVersion(cfg.Vault.Version)),
+		client:            kvclient.NewAdditiveClient(c.Logical(), cfg.Vault.MountPath, kvclient.WithVersion(cfg.Vault.Version)),
 		defaultParentPath: cfg.DefaultScope,
 	}, nil
 }
 
-// ReadKeyValues reads and returns key value pairs for a given Vault Secret.
+// ReadKeyValues reads and returns key value pairs for a given Vault KVSecret.
 func (ss *SecretStore) ReadKeyValues(_ context.Context, i store.Secret) (store.KeyValues, error) {
-	s := &KVSecret{}
-	if err := ss.client.Get(ss.pathForSecretInstance(i), s); resource.Ignore(isNotFound, err) != nil {
+	s := &kvclient.KVSecret{}
+	if err := ss.client.Get(ss.pathForSecretInstance(i), s); resource.Ignore(kvclient.IsNotFound, err) != nil {
 		return nil, errors.Wrap(err, errGet)
 	}
 	kv := make(store.KeyValues, len(s.Data))
@@ -104,14 +104,15 @@ func (ss *SecretStore) ReadKeyValues(_ context.Context, i store.Secret) (store.K
 	return kv, nil
 }
 
-// WriteKeyValues writes key value pairs to a given Vault Secret.
+// WriteKeyValues writes key value pairs to a given Vault KVSecret.
 func (ss *SecretStore) WriteKeyValues(_ context.Context, i store.Secret, kv store.KeyValues) error {
 	data := make(map[string]interface{}, len(kv))
 	for k, v := range kv {
 		data[k] = string(v)
 	}
 
-	kvSecret := &KVSecret{Data: data}
+	kvSecret := &kvclient.KVSecret{}
+	kvSecret.Data = data
 	if i.Metadata != nil && len(i.Metadata.Labels) > 0 {
 		kvSecret.CustomMeta = make(map[string]interface{}, len(i.Metadata.Labels))
 		for k, v := range i.Metadata.Labels {
@@ -122,19 +123,36 @@ func (ss *SecretStore) WriteKeyValues(_ context.Context, i store.Secret, kv stor
 	return errors.Wrap(ss.client.Apply(ss.pathForSecretInstance(i), kvSecret), errApply)
 }
 
-// DeleteKeyValues delete key value pairs from a given Vault Secret.
+// DeleteKeyValues delete key value pairs from a given Vault KVSecret.
 // If no kv specified, the whole secret instance is deleted.
 // If kv specified, those would be deleted and secret instance will be deleted
 // only if there is no Data left.
-func (ss *SecretStore) DeleteKeyValues(_ context.Context, i store.Secret, _ store.KeyValues) error {
-	// TODO(turkenh): Handle deletion of partial kv, currently we delete the
-	//  whole secret.
-	return errors.Wrap(ss.client.Delete(ss.pathForSecretInstance(i)), errDelete)
+func (ss *SecretStore) DeleteKeyValues(_ context.Context, i store.Secret, kv store.KeyValues) error {
+	s := &kvclient.KVSecret{}
+	err := ss.client.Get(ss.pathForSecretInstance(i), s)
+	if kvclient.IsNotFound(err) {
+		// KVSecret already deleted, nothing to do.
+		return nil
+	}
+	if err != nil {
+		return errors.Wrap(err, errGet)
+	}
+	for k := range kv {
+		delete(s.Data, k)
+	}
+	if len(kv) == 0 || len(s.Data) == 0 {
+		// KVSecret is deleted only if:
+		// - No kv to delete specified as input
+		// - No data left in the secret
+		return errors.Wrap(ss.client.Delete(ss.pathForSecretInstance(i)), errDelete)
+	}
+	// If there are still keys left, update the secret with the remaining.
+	return errors.Wrap(ss.client.Apply(ss.pathForSecretInstance(i), s), errApply)
 }
 
 func (ss *SecretStore) pathForSecretInstance(i store.Secret) string {
 	if i.Scope != "" {
-		return filepath.Clean(filepath.Join(i.Scope, i.Name))
+		return filepath.Join(i.Scope, i.Name)
 	}
-	return filepath.Clean(filepath.Join(ss.defaultParentPath, i.Name))
+	return filepath.Join(ss.defaultParentPath, i.Name)
 }
